requirements.txt
certifi @ file:///croot/certifi_1671487769961/work/certifi
cycler==0.11.0
easydict==1.10
fonttools==4.38.0
glob2==0.7
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.7.0.68
packaging==22.0
Pillow==9.4.0
protobuf==3.20.1
pyaml==21.10.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
tensorboardX==2.5.1
torch==1.3.1
torchvision==0.4.2
tqdm==4.64.1
typing_extensions==4.4.0
